requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
opencv-python>=4.5.0
pillow>=8.0.0
streamlit>=1.8.0
mediapipe>=0.8.10